First commit of sdl-ttf-opengl-cpp
This is a project that provides TTF support to my sdl-opengl-cpp
project.

It's not perfect, there are a lot of things that need to be fixed up.
But it is eating my own dog food and showing how to use my first CMake
package from another package.

It's already revealed some better practices around multiple library
targets for exception and no-exception versions.